Extract import logic from index to it's own file
